refactor(mobile): extract CrashReportingToggle widget + guard re-entrant toggles
Move the private _CrashReportingToggle out of user_profile_screen.dart into a
public CrashReportingToggle (lib/profile/widgets/crash_reporting_toggle.dart)
with injectable readEnabled/setEnabled seams (default to the real functions) so
it's widget-testable without the native Sentry SDK. Adds an in-flight guard:
the switch is disabled (onChanged null) while a toggle awaits, so a rapid
double-tap can't start a second concurrent SDK init/close. (CodeRabbit #645)
